+++ changelog.md
@@ -1,2 +1,6 @@
+# v2.0.0
+- Use Geode's custom progress bar utilities (thanks me)
+- Hook all progress bars found in-game
+
 # v1.0.0
 - Initial release